Move base class to module
