+++ qiskit_ibm_runtime/decoders/noise_learner.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from ..utils.noise_learner_result import LayerError, NoiseLearnerResult, PauliLindbladError
-from ..utils.result_decoder import ResultDecoder
+from .result_decoder import ResultDecoder
 from ..decoders.noise_learner_v3.decoder import NoiseLearnerV3ResultDecoder
 
 if TYPE_CHECKING:
