+++ qiskit_ibm_runtime/decoders/quantum_program/decoder.py
@@ -27,7 +27,7 @@ from ibm_quantum_schemas.executor.version_1_0 import (
     QuantumProgramResultModel as QuantumProgramResultModel_1_0,
 )
 
-from ...utils.result_decoder import ResultDecoder
+from ..result_decoder import ResultDecoder
 from ..executor_sampler.post_processor_v0_1 import (
     sampler_v2_post_processor_v0_1,
 )
